pathfinding game object in working state
returns the next tile to go to, to improve it i could make it only call the function when its stood on its destination, rather than calling the function for the same result every tick, when its still on the same tile as the last call. the only time it needs updating is when the enemy has reached the target tile or when the player has changed tile. otherwise the return of the function will not change and calling it multiple times is going to negatively affect performance